requirements.txt
asgiref==3.4.1
beautifulsoup4==4.10.0
Django==3.2.8
django-admin-thumbnails==0.2.6
django-bootstrap4==3.0.1
django-fontawesome-5==1.0.18
gunicorn==20.1.0
Pillow==8.3.2
pytz==2021.3
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.2
whitenoise==5.3.0
